update mqtt python scripts
